Set context when creating kind cluster
